Move logger mock to public directory
