+++ pkg/secretless/log/mock/logger.go
